fix: harden network id generation and tests

+++ engine/src/net/component/object.rs
@@ -6,7 +6,7 @@ use crate::utils::{ReflectTypeUuidDynamic, TypeUuid};
 use renet::ClientId;
 use serde::{Deserialize, Serialize};
 
-pub type NetworkObjectId = u32;
+pub type NetworkObjectId = u64;
 
 #[derive(Default, TypeUuid, Serialize, Deserialize, Component, Reflect)]
 #[uuid = "adca3c70-5d3d-4b32-83cc-7bdf04a8358a"]
